chore(dotnet): refactor parsing function into util
